Reliable support for conditional imports
I was struggling with dynamic imports for a while before I realized I
was actually working on the halting problem. This is hinted at in the
top-level await proposal:
https://github.com/tc39/proposal-top-level-await#rejected-deadlock-prevention-mechanisms

Instead, we treat resolved dynamic imports basically as static imports,
and they will participate in the cyclic resolution algorithm. Since
dynamic imports in practice tend to just be well-behaved conditional or
lazy imports this works out well.

This are probably some races here if a new dynamic import is dispatched
while an update is processing but haven't actually run into it myself.