Improve diagnostics
Unaccepted updates will print a tree showing how we got there.
Unresolvable specifiers which poison an `accept` group will also be
logged. Finally, logged module paths are printed relative to the current
working directory.

Also fixes an issue in the transformer which failed to detect top-level
await when nested in a non-function block.