Fix self-accept documentation
Note: In Webpack the callback is `onError`, but Vite and esm-hmr specify
that it is `onUpdate`. Vite does not implement live bindings, so a
handler is required. Webpack does not do error propagation very well, so
an error handler is required. dynohot propagates errors and implements
live bindings so the callback is kind of vestigial. It is best to simply
not use this parameter.